+++ pytsite/console/_option.py
@@ -1,7 +1,7 @@
 """
 """
 from typing import Any as _Any, List as _List
-from pytsite import formatter as _formatter, validation as _validation, util as _util
+from pytsite import formatters as _formatter, validation as _validation, util as _util
 from . import _error
 
 __author__ = 'Alexander Shepetko'